More docs for names and space
