+++ src/main/scala/parsley/token/descriptions/NameDesc.scala
@@ -5,11 +5,25 @@ package parsley.token.descriptions
 
 import parsley.token.predicate.{CharPredicate, NotRequired}
 
+/** The class describes how name-like things are described lexically.
+  *
+  * @param identifierStart what characters may start an identifier?
+  * @param identifierLetter what characters may continue an identifier?
+  * @param operatorStart what characters may start a user-defined operator?
+  * @param operatorLetter what characters may continue a user-defined operator?
+  * @since 4.0.0
+  */
 final case class NameDesc (identifierStart: CharPredicate,
                            identifierLetter: CharPredicate,
                            operatorStart: CharPredicate,
                            operatorLetter: CharPredicate)
 
+/** This object contains any preconfigured name definitions.
+  * @since 4.0.0
+  */
 object NameDesc {
+    /** Plain description of names, where neither identifiers nor operators are required.
+      * @since 4.0.0
+      */
     val plain = NameDesc(NotRequired, NotRequired, NotRequired, NotRequired)
 }
